fix wrong folder names
