+++ examples/vectoradd_jax/example_abstract_inputs.json
