+++ examples/vectoradd_jax/example_jvp_inputs.json
@@ -18,7 +18,7 @@
     },
     "jvp_inputs": [
         "a.v",
-        "b.v"
+        "a.s"
     ],
     "jvp_outputs": [
         "vector_add.result",
@@ -30,10 +30,6 @@
                 5.0,
                 2.0
             ],
-        "b.v": [
-                1.0,
-                0.0,
-                3.0
-        ]
+        "a.s": 1.0
     }
 }
