+++ examples/vectoradd_torch/example_jvp_inputs.json
@@ -18,7 +18,7 @@
     },
     "jvp_inputs": [
         "a.v",
-        "a.s"
+        "b.v"
     ],
     "jvp_outputs": [
         "vector_add.result",
@@ -30,6 +30,10 @@
                 5.0,
                 2.0
             ],
-        "a.s": 1.0
+        "b.v": [
+                1.0,
+                0.0,
+                3.0
+        ]
     }
 }
